refactor: move date redux files to store
