+++ src/actions/types.js
@@ -17,6 +17,3 @@ export const REMOVE_ITEM_ERROR = 'REMOVE_ITEM_ERROR'
 export const UPDATE_ITEM_REQUEST = 'UPDATE_ITEM_REQUEST'
 export const UPDATE_ITEM_SUCCESS = 'UPDATE_ITEM_SUCCESS'
 export const UPDATE_ITEM_ERROR = 'UPDATE_ITEM_ERROR'
-
-export const NEXT_WEEK = 'NEXT_WEEK'
-export const PREVIOUS_WEEK = 'PREVIOUS_WEEK'
