+++ src/reducers/index.js
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux'
 import { routerReducer } from 'react-router-redux'
 import loading from './loadingReducer'
 import error from './errorReducer'
-import date from './dateReducer'
+import date from '../store/date/dateReducer'
 import items from './itemsReducer'
 import authentication from '../store/auth/authReducer'
 import expense from './expenseReducer'
