+++ src/store/date/dateActions.js
@@ -1,4 +1,5 @@
-import { NEXT_WEEK, PREVIOUS_WEEK } from './types'
+export const NEXT_WEEK = 'NEXT_WEEK'
+export const PREVIOUS_WEEK = 'PREVIOUS_WEEK'
 
 export const nextWeek = payload => ({
   type: NEXT_WEEK,
